requirements.txt
fastapi[standard]~=0.115.14
cryptography==45.0.4
uvicorn~=0.35.0
sqlmodel~=0.0.24
onepassword-sdk~=0.3.1
SQLAlchemy~=2.0.41
fabric~=3.2.2